# api/api.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: ElektroGo
  description: "Documentació OpenAPI per a l'API REST desenvolupada a partir de l'aplicació **ElektroGo**. <br> Per tal de poder executar les operacions cal estar conectat a la VPN de la UPC. Totes les operacions requereixen d'una api_key. En cas de no tenir-ne la podeu solicitar a qualsevol membre de l'equip."
servers:
  - url: 'http://10.4.41.58:8080/'
paths:
...
  /drivers/{username}:
    get:
      tags:
      - Drivers
      summary: Obté un conductor.
      operationId: getDriver
      parameters:
      - name: username
        in: path
        description: Username del conductor
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object 
                $ref: '#/components/schemas/Driver'
        432:
          description: 'Error: Driver Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error432'
    post:
      tags:
      - Drivers
      summary: Crea un conductor.
      operationId: createDriver
      parameters:
      - name: username
        in: path
        description: Username del conductor
        required: true
        schema:
          type: string
      responses:
        201:
          description: successful operation
        439:
          description: 'Error: Driver Alredy Exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error439'
    delete:
      tags:
      - Drivers
      summary: Esborra un conductor.
      operationId: deleteDriver
      parameters:
      - name: username
        in: path
        description: Username del conductor
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
        432:
          description: 'Error: Driver Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error432'
  /drivers/{username}/verify:
    put:
      tags:
      - Drivers
      summary: Verifica un conductor.
      operationId: verifyDriver
      parameters:
      - name: username
        in: path
        description: Username del conductor
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
        432:
          description: 'Error: Driver Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error432'
...
components:
  schemas:
...
    Driver:
      type: object
      properties:
        username:
          type: string
          example: "samito"
        status:
          type: string
          example: "verified"
...
    ErrorResultElektroGo:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        path:
          type: string
  examples:
...
    error432:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 432
        error: "Http Status 432"
        path: "/drivers"
...
    error439:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 439
        error: "Http Status 439"
        path: "/drivers/samito"
...